--- src/main/java/com/gabrielosorio/gestor_inteligente/repository/CategoryRepository.java
@@ -1,6 +1,0 @@
-package com.gabrielosorio.gestor_inteligente.repository;
-
-import com.gabrielosorio.gestor_inteligente.model.Category;
-
-public class CategoryRepository extends Repository<Category> {
-}
