--- src/main/java/com/gabrielosorio/gestor_inteligente/repository/PaymentRepository.java
@@ -1,6 +1,0 @@
-package com.gabrielosorio.gestor_inteligente.repository;
-import com.gabrielosorio.gestor_inteligente.model.Payment;
-
-public class PaymentRepository extends Repository<Payment> {
-}
-
