--- src/main/java/com/gabrielosorio/gestor_inteligente/repository/SaleRepository.java
@@ -1,6 +1,0 @@
-package com.gabrielosorio.gestor_inteligente.repository;
-import com.gabrielosorio.gestor_inteligente.model.Sale;
-
-public class SaleRepository extends Repository<Sale> {
-
-}
