+++ src/main/java/com/gabrielosorio/gestor_inteligente/repository/base/Repository.java
@@ -1,4 +1,4 @@
-package com.gabrielosorio.gestor_inteligente.repository;
+package com.gabrielosorio.gestor_inteligente.repository.base;
 
 import com.gabrielosorio.gestor_inteligente.repository.specification.Specification;
 import com.gabrielosorio.gestor_inteligente.repository.strategy.base.BatchInsertable;
