+++ src/main/java/com/gabrielosorio/gestor_inteligente/repository/impl/CategoryRepository.java
@@ -1,0 +1,7 @@
+package com.gabrielosorio.gestor_inteligente.repository.impl;
+
+import com.gabrielosorio.gestor_inteligente.model.Category;
+import com.gabrielosorio.gestor_inteligente.repository.base.Repository;
+
+public class CategoryRepository extends Repository<Category> {
+}
