+++ src/main/java/com/gabrielosorio/gestor_inteligente/repository/impl/CheckoutMovementRepository.java
@@ -1,9 +1,10 @@
-package com.gabrielosorio.gestor_inteligente.repository;
+package com.gabrielosorio.gestor_inteligente.repository.impl;
 import com.gabrielosorio.gestor_inteligente.model.CheckoutMovement;
+import com.gabrielosorio.gestor_inteligente.repository.base.Repository;
 import com.gabrielosorio.gestor_inteligente.repository.specification.FindCheckoutMovementByCheckoutId;
 import java.util.List;
 
-public class CheckoutMovementRepository extends Repository<CheckoutMovement>{
+public class CheckoutMovementRepository extends Repository<CheckoutMovement> {
 
     public List<CheckoutMovement> findByCheckoutId(long checkoutId) {
         return findBySpecification(new FindCheckoutMovementByCheckoutId(checkoutId));
