+++ src/main/java/com/gabrielosorio/gestor_inteligente/repository/impl/CheckoutMovementTypeRepository.java
@@ -1,5 +1,6 @@
-package com.gabrielosorio.gestor_inteligente.repository;
+package com.gabrielosorio.gestor_inteligente.repository.impl;
 import com.gabrielosorio.gestor_inteligente.model.CheckoutMovementType;
+import com.gabrielosorio.gestor_inteligente.repository.base.Repository;
 
 public class CheckoutMovementTypeRepository extends Repository<CheckoutMovementType> {
 }
