+++ src/main/java/com/gabrielosorio/gestor_inteligente/repository/impl/PaymentRepository.java
@@ -1,0 +1,7 @@
+package com.gabrielosorio.gestor_inteligente.repository.impl;
+import com.gabrielosorio.gestor_inteligente.model.Payment;
+import com.gabrielosorio.gestor_inteligente.repository.base.Repository;
+
+public class PaymentRepository extends Repository<Payment> {
+}
+
