+++ src/main/java/com/gabrielosorio/gestor_inteligente/repository/impl/ProductRepository.java
@@ -1,6 +1,7 @@
-package com.gabrielosorio.gestor_inteligente.repository;
+package com.gabrielosorio.gestor_inteligente.repository.impl;
 
 import com.gabrielosorio.gestor_inteligente.model.Product;
+import com.gabrielosorio.gestor_inteligente.repository.base.Repository;
 import com.gabrielosorio.gestor_inteligente.repository.strategy.base.ProductRepositoryStrategy;
 
 public class ProductRepository extends Repository<Product> {
