+++ src/main/java/com/gabrielosorio/gestor_inteligente/repository/impl/SaleCheckoutMovementRepository.java
@@ -1,5 +1,6 @@
-package com.gabrielosorio.gestor_inteligente.repository;
+package com.gabrielosorio.gestor_inteligente.repository.impl;
 import com.gabrielosorio.gestor_inteligente.model.SaleCheckoutMovement;
+import com.gabrielosorio.gestor_inteligente.repository.base.Repository;
 
 public class SaleCheckoutMovementRepository extends Repository<SaleCheckoutMovement> {
