+++ src/main/java/com/gabrielosorio/gestor_inteligente/repository/impl/SalePaymentRepository.java
@@ -1,6 +1,7 @@
-package com.gabrielosorio.gestor_inteligente.repository;
+package com.gabrielosorio.gestor_inteligente.repository.impl;
 
 import com.gabrielosorio.gestor_inteligente.model.SalePayment;
+import com.gabrielosorio.gestor_inteligente.repository.base.Repository;
 
 public class SalePaymentRepository extends Repository<SalePayment> {
