+++ src/main/java/com/gabrielosorio/gestor_inteligente/repository/impl/SaleProductRepository.java
@@ -1,6 +1,7 @@
-package com.gabrielosorio.gestor_inteligente.repository;
+package com.gabrielosorio.gestor_inteligente.repository.impl;
 
 import com.gabrielosorio.gestor_inteligente.model.SaleProduct;
+import com.gabrielosorio.gestor_inteligente.repository.base.Repository;
 
 public class SaleProductRepository extends Repository<SaleProduct> {
