+++ src/main/java/com/gabrielosorio/gestor_inteligente/repository/impl/SaleRepository.java
@@ -1,0 +1,7 @@
+package com.gabrielosorio.gestor_inteligente.repository.impl;
+import com.gabrielosorio.gestor_inteligente.model.Sale;
+import com.gabrielosorio.gestor_inteligente.repository.base.Repository;
+
+public class SaleRepository extends Repository<Sale> {
+
+}
