+++ src/main/java/com/gabrielosorio/gestor_inteligente/service/impl/SaleProductServiceImpl.java
@@ -1,7 +1,7 @@
 package com.gabrielosorio.gestor_inteligente.service.impl;
 
 import com.gabrielosorio.gestor_inteligente.model.SaleProduct;
-import com.gabrielosorio.gestor_inteligente.repository.Repository;
+import com.gabrielosorio.gestor_inteligente.repository.base.Repository;
 
 import com.gabrielosorio.gestor_inteligente.service.AbstractTransactionalService;
 import com.gabrielosorio.gestor_inteligente.service.base.SaleProductService;
