+++ src/main/java/com/gabrielosorio/gestor_inteligente/view/ProductManagerController.java
@@ -3,8 +3,8 @@ import com.gabrielosorio.gestor_inteligente.GestorInteligenteApp;
 import com.gabrielosorio.gestor_inteligente.config.ConnectionFactory;
 import com.gabrielosorio.gestor_inteligente.model.Product;
 import com.gabrielosorio.gestor_inteligente.model.Stock;
-import com.gabrielosorio.gestor_inteligente.repository.ProductRepository;
-import com.gabrielosorio.gestor_inteligente.repository.storage.PSQLProductStrategy;
+import com.gabrielosorio.gestor_inteligente.repository.impl.ProductRepository;
+import com.gabrielosorio.gestor_inteligente.repository.strategy.psql.PSQLProductStrategy;
 import com.gabrielosorio.gestor_inteligente.service.base.ProductService;
 import com.gabrielosorio.gestor_inteligente.service.impl.ProductServiceImpl;
 import com.gabrielosorio.gestor_inteligente.utils.TextFieldUtils;
