+++ src/main/java/com/gabrielosorio/gestor_inteligente/view/ProductTbViewController.java
@@ -3,8 +3,8 @@ package com.gabrielosorio.gestor_inteligente.view;
 import com.gabrielosorio.gestor_inteligente.config.ConnectionFactory;
 import com.gabrielosorio.gestor_inteligente.model.Category;
 import com.gabrielosorio.gestor_inteligente.model.Product;
-import com.gabrielosorio.gestor_inteligente.repository.ProductRepository;
-import com.gabrielosorio.gestor_inteligente.repository.storage.PSQLProductStrategy;
+import com.gabrielosorio.gestor_inteligente.repository.impl.ProductRepository;
+import com.gabrielosorio.gestor_inteligente.repository.strategy.psql.PSQLProductStrategy;
 import com.gabrielosorio.gestor_inteligente.service.base.ProductService;
 import com.gabrielosorio.gestor_inteligente.service.impl.ProductServiceImpl;
 import com.gabrielosorio.gestor_inteligente.utils.TableViewUtils;
